Clarify "community staff" in CoC
This is a first pass at a resolution to #33.

I've added email links for a few people that already had their email
addresses publicly available and that I felt confident would be happy to
be on this contact list. We can definitely add more contacts if there
are others who would be willing to be contacts.