make rss great again
I am a user of rss feed aggregators and find the current rss feed not as
useful as it could be. You can checkout the current state by looking at
typelevel.org/blog/feed.rss on feedbucket.com. The list of blog posts is
limited to 10 entries and the descriptions are very terse.

The terseness, I think, is a bug with the `{% if post.excerpt %}` line.
At some point Jekyll decided to auto-generate excerpts if one is not
given in the YAML header. So I think this `if` statement is always true.

I greped the posts directory and found no use of the `excerpt` variable
so I made the decision to just remove the use of the variable in the
feed.rss file.

The end result is a useful rss feed. It can be previewed at
feedbucket.com using this link
https://gist.githubusercontent.com/jarrodwb/003ea9995134bde814e6d3545529e483/raw/bbc5d244c734f55fce77dc5b08fc8140be32abea/rss-fixes.rss.

The formatting is beautiful on feedbin.com. This is the aggregator I use.