Add Product with Serializable blog post
This is a document that I've had sitting on my laptop for a while. I
figured that I might as well put it in a public place, and the typelevel
blog is the first place that came to mind.

I now see that Noel Welsh has published something similar [here](https://underscore.io/blog/posts/2015/06/04/more-on-sealed.html).
This goes into a bit more detail with type LUBs, which may be helpful to
someone?