Add a blog post about variance and monad transformers
Every once in a while [people](https://github.com/typelevel/cats/issues/556) [ask](https://github.com/typelevel/cats/issues/2310) [about](https://github.com/typelevel/cats/issues/2538) this.

I thought that it would be nice to provide a more detailed answer in a
single place so we can link to it from the Cats FAQ.

Note: I had to add the `-Ypartial-unification` flag for one of the code
samples to compile. It's quite possible that I should be doing this a
different way than I did.